[TwigBundle] Move the code filters to a dedicated extensions
A dedicated extension now exists for the code-related filters for Twig.
The dependency to service_container was also removed, to use CodeHelper, instead
