+++ src/Symfony/Bundle/TwigBundle/Extension/TemplatingExtension.php
@@ -27,23 +27,6 @@ class TemplatingExtension extends \Twig_Extension
     }
 
     /**
-     * {@inheritdoc}
-     */
-    public function getFilters()
-    {
-        return array(
-            'abbr_class'            => new \Twig_Filter_Method($this, 'abbrClass', array('is_safe' => array('html'))),
-            'abbr_method'           => new \Twig_Filter_Method($this, 'abbrMethod', array('is_safe' => array('html'))),
-            'format_args'           => new \Twig_Filter_Method($this, 'formatArgs', array('is_safe' => array('html'))),
-            'format_args_as_text'   => new \Twig_Filter_Method($this, 'formatArgsAsText'),
-            'file_excerpt'          => new \Twig_Filter_Method($this, 'fileExcerpt', array('is_safe' => array('html'))),
-            'format_file'           => new \Twig_Filter_Method($this, 'formatFile', array('is_safe' => array('html'))),
-            'format_file_from_text' => new \Twig_Filter_Method($this, 'formatFileFromText', array('is_safe' => array('html'))),
-            'file_link'             => new \Twig_Filter_Method($this, 'getFileLink', array('is_safe' => array('html'))),
-        );
-    }
-
-    /**
      * Returns a list of functions to add to the existing list.
      *
      * @return array An array of functions
@@ -80,46 +63,6 @@ class TemplatingExtension extends \Twig_Extension
     public function getAssetsVersion($packageName = null)
     {
         return $this->container->get('templating.helper.assets')->getVersion($packageName);
-    }
-
-    public function abbrClass($class)
-    {
-        return $this->container->get('templating.helper.code')->abbrClass($class);
-    }
-
-    public function abbrMethod($method)
-    {
-        return $this->container->get('templating.helper.code')->abbrMethod($method);
-    }
-
-    public function formatArgs($args)
-    {
-        return $this->container->get('templating.helper.code')->formatArgs($args);
-    }
-
-    public function formatArgsAsText($args)
-    {
-        return $this->container->get('templating.helper.code')->formatArgsAsText($args);
-    }
-
-    public function fileExcerpt($file, $line)
-    {
-        return $this->container->get('templating.helper.code')->fileExcerpt($file, $line);
-    }
-
-    public function formatFile($file, $line, $text = null)
-    {
-        return $this->container->get('templating.helper.code')->formatFile($file, $line, $text);
-    }
-
-    public function getFileLink($file, $line)
-    {
-        return $this->container->get('templating.helper.code')->getFileLink($file, $line);
-    }
-
-    public function formatFileFromText($text)
-    {
-        return $this->container->get('templating.helper.code')->formatFileFromText($text);
     }
 
     /**
