+++ src/Symfony/Bundle/TwigBundle/Resources/config/twig.xml
@@ -48,6 +48,11 @@
             <argument type="service" id="templating.helper.actions" />
         </service>
 
+        <service id="twig.extension.code" class="Symfony\Bundle\TwigBundle\Extension\CodeExtension" public="false">
+            <tag name="twig.extension" />
+            <argument type="service" id="templating.helper.code" />
+        </service>
+
         <service id="twig.extension.routing" class="Symfony\Bridge\Twig\Extension\RoutingExtension" public="false">
             <tag name="twig.extension" />
             <argument type="service" id="router" />
